+++ lib/formular/element/bootstrap4/custom_control.rb
@@ -1,5 +1,5 @@
 require 'formular/elements'
-require 'formular/element/modules/wrapped_control'
+require 'formular/element/modules/wrapped'
 require 'formular/element/bootstrap4'
 require 'formular/element/bootstrap3'
 require 'formular/element/module'
@@ -10,8 +10,9 @@ module Formular
       module CustomControl
         module CustomCheckable
           include Formular::Element::Module
-          include Formular::Element::Modules::WrappedControl
+          include Formular::Element::Modules::Wrapped
           set_default :class, ['custom-control-input']
+          set_default :label_options, { class: ['form-control-label'] }
 
           html(:checkable_label) do |input|
             Formular::Element::Label.(input.label_options) do
@@ -57,7 +58,6 @@ module Formular
             html { closed_start_tag }
 
             set_default :control_label_options, { class: ['custom-control custom-checkbox'] }
-            set_default :value, '1' # instead of reader value
           end # class Checkbox
 
           class CustomRadio < Formular::Element::Radio
@@ -71,13 +71,15 @@ module Formular
 
           class CustomSelect < Formular::Element::Bootstrap3::Select
             set_default :class, ['custom-select']
+            set_default :label_options, { class: ['form-control-label'] }
           end # class CustomSelect
 
           class CustomFile < Formular::Element::Input
-            include Formular::Element::Modules::WrappedControl
+            include Formular::Element::Modules::Wrapped
             include Formular::Element::Bootstrap3::ColumnControl
 
             set_default :class, ['custom-file-input']
+            set_default :label_options, { class: ['form-control-label'] }
             set_default :type, 'file'
 
             rename_html_context(:default, :control)
@@ -93,6 +95,7 @@ module Formular
 
         class CustomSelect < Inline::CustomSelect
           rename_html_context(:default, :input)
+          set_default :label_options, { class: ['form-control-label'] }
 
           html do |input|
             Formular::Element::Div.(content: input.to_html(context: :input))
@@ -101,6 +104,7 @@ module Formular
 
         class CustomFile < Inline::CustomFile
           rename_html_context(:default, :input)
+          set_default :label_options, { class: ['form-control-label'] }
 
           html do |input|
             Formular::Element::Div.(content: input.to_html(context: :input))