+++ test/builders/bootstrap3_test.rb
@@ -60,7 +60,7 @@ describe 'Bootstrap3 builders' do
         ).to_s
         html << form.end
 
-        html.must_equal %(<form action="/questions/13" method="post" accept-charset=\"utf-8\"><input name=\"utf8\" type=\"hidden\" value=\"✓\"/><div class="form-group"><label for="comment_body" class="control-label">Some fancy label</label><textarea name="comment[body]" id="comment_body" class="form-control">Something exciting</textarea></div><div class="form-group"><label for=\"comment_public\" class="control-label">Another fancy label</label><select name="comment[public]" id="comment_public" class="form-control"><option value="1">Option 1</option><option value="2">Option 2</option></select></div><div class="form-group"><label for="comment_body" class="control-label">Oh look, here's another fancy label</label><input name="comment[body]" id="comment_body" value="Something exciting" type="text" class="form-control"/></div></form>)
+        html.must_equal %(<form action="/questions/13" method="post" accept-charset=\"utf-8\"><input name=\"utf8\" type=\"hidden\" value=\"✓\"/><div class="form-group"><label for="comment_body" class="control-label">Some fancy label</label><textarea name="comment[body]" id="comment_body" class="form-control">Something exciting</textarea></div><div class="form-group"><label for=\"comment_public\" class="control-label">Another fancy label</label><select name="comment[public]" id="comment_public" class="form-control"><option value="1">Option 1</option><option value="2">Option 2</option></select></div><div class="form-group"><label for="comment_body" class="control-label">Oh look, here&#39;s another fancy label</label><input name="comment[body]" id="comment_body" value="Something exciting" type="text" class="form-control"/></div></form>)
       end
     end
 
@@ -97,7 +97,7 @@ describe 'Bootstrap3 builders' do
         html.must_equal %(<form action="/questions/13" method="post" accept-charset="utf-8" class="form-horizontal"><input name=\"utf8\" type=\"hidden\" value=\"✓\"/><div class="form-group"><label for="comment_body" class="col-sm-2 control-label">Body</label><div class="col-sm-10"><input name="comment[body]" id="comment_body" value="Something exciting" type="text" class="form-control"/></div></div></form>)
       end
 
-      it 'wrapped_control renders correctly with hint & error' do
+      it 'wrapped renders correctly with hint & error' do
         form = builder.form(action: '/questions/13') do |f|
           f.input(:body, label: 'Body', hint: 'some handy hint', error: 'Something went wrong').to_s
         end