+++ test/test_helper.rb
@@ -9,7 +9,7 @@ Owner   = Struct.new(:id, :name, :email, :errors)
 COLLECTION_ARRAY = [['Option 1', 1], ['Option 2', 2]]
 
 # Manu test
-User = Struct.new(:id, :firstname, :lastname, :email, :password, :confirm_password)
+User = Struct.new(:id, :firstname, :lastname, :gender, :dob, :email, :password, :confirm_password)
 Post = Struct.new(:id, :title, :subtitle, :body, :image)
 
 def generate_xml(string)